Sheniece's Chess Project Contribution Code
This file solely contains my edited code; the code begins with the headers, then moves onto the source code, and then finally concludes with the design code. 
It was made in QT creator and incorporates a SQLite database that is linked to the code, and displays evidence of its ongoing connection within the code when it is executed.